# src/main/resources/static/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Whakaoko",
    "description": "A service for aggregating 3rd party content from a variety of sources (RSS / Twitter) into an easy to consume local JSON or RSS feed.",
    "version": "1.0.0"
  },
  "paths": {
    "/channels": {
      "post": {
        "summary": "Create a channel",
        "tags": [
          "Channels"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateChannelRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "The created channel",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Channel"
                }
              }
            }
          }
        }
      }
    },
    "/{username}/channels": {
      "get": {
        "summary": "List a user's channels",
        "tags": [
          "Channels"
        ],
        "parameters": [
          {
            "name": "username",
            "in": "path",
            "description": "A username",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "A list of channels",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Channel"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/channels/{id}": {
      "get": {
        "summary": "Show single channel",
        "tags": [
          "Channels"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Channel id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "A single channel",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Channel"
                }
              }
            }
          }
        }
      }
    },
    "/channels/{id}/items": {
      "get": {
        "summary": "List feed items for this channel.",
        "tags": [
          "Channels"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Channel id",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "pageSize",
            "in": "query",
            "required": false,
            "description": "Optional page size",
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "A list of feed items",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/FeedItem"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/channels/{id}/subscriptions": {
      "get": {
        "summary": "List the subscriptions which make up this channel.",
        "tags": [
          "Channels"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Channel id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "A list of subscriptions",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Subscription"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/subscriptions": {
      "post": {
        "summary": "Create subscription",
        "tags": [
          "Subscriptions"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SubscriptionCreateRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "The created subscription",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Subscription"
                }
              }
            }
          }
        }
      }
    },
    "/subscriptions/{id}": {
      "get": {
        "summary": "Show subscription",
        "tags": [
          "Subscriptions"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Subscription id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "A single subscription",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Subscription"
                }
              }
            }
          }
        }
      },
      "put": {
        "summary": "Update a subscription",
        "tags": [
          "Subscriptions"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Subscription id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateSubscriptionRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "The updated subscription",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Subscription"
                }
              }
            }
          }
        }
      }
    },
    "/subscriptions/{id}/items": {
      "get": {
        "summary": "List feed items for this subscription.",
        "tags": [
          "Subscriptions"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Subscription id",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "pageSize",
            "in": "query",
            "required": false,
            "description": "Optional page size",
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "A list of feed items",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/FeedItem"
                  }
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "Channel": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "username": {
            "type": "string"
          }
        }
      },
      "Category": {
        "type": "object",
        "properties": {
          "value": {
            "type": "string"
          }
        }
      },
      "FeedItem": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string"
          },
          "url": {
            "type": "string"
          },
          "body": {
            "type": "string"
          },
          "subscriptionId": {
            "type": "string"
          },
          "channelId": {
            "type": "string"
          },
          "date": {
            "type": "string",
            "format": "date-time"
          },
          "categories": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Category"
            }
          }
        }
      },
      "CreateChannelRequest": {
        "type": "object",
        "required": [
          "name"
        ],
        "properties": {
          "name": {
            "type": "string"
          }
        }
      },
      "UpdateSubscriptionRequest": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          }
        }
      },
      "SubscriptionCreateRequest": {
        "type": "object",
        "required": [
          "channel",
          "url"
        ],
        "properties": {
          "name": {
            "type": "string"
          },
          "url": {
            "type": "string",
            "description": "The URL of an RSS feed."
          },
          "channel": {
            "type": "string",
            "description": "The id of the channel to add this feed to."
          }
        }
      },
      "Subscription": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "url": {
            "type": "string"
          },
          "username": {
            "type": "string"
          },
          "latestItemDate": {
            "type": "string",
            "format": "date-time"
          },
          "lastRead": {
            "type": "string",
            "format": "date-time"
          },
          "etag": {
            "type": "string"
          },
          "classification": {
            "type": "string"
          }
        }
      }
    }
  }
}